Added support for custom Markdown block parsing. Also separated source code into more files for easier comprehension.
