+++ super_editor_markdown/lib/super_editor_markdown.dart
@@ -1,1 +1,3 @@
-export 'src/markdown.dart';
+export 'src/document_to_markdown_serializer.dart';
+export 'src/markdown_to_document_parsing.dart';
+export 'src/super_editor_syntax.dart';
