+++ super_editor_markdown/test/attributed_text_markdown_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:super_editor/super_editor.dart';
-import 'package:super_editor_markdown/src/markdown.dart';
+import 'package:super_editor_markdown/src/document_to_markdown_serializer.dart';
 
 void main() {
   group("AttributedText markdown serializes", () {
